Документация, диаграммы, E2E-сценарий и CI
- 5 code-generated диаграмм PlantUML: CJM пользователя и заведения,
  конечный автомат заказа, C4 Level 2, ERD (рендер через встроенный
  Smetana — graphviz не требуется)
- 4 ADR с разбором альтернатив: PostgreSQL как источник правды,
  версионирование меню и снапшоты, transactional outbox, аутентификация
  партнёров
- scripts/smoke.sh: 12 шагов сквозного сценария с проверкой HTTP-статусов
  и доменных кодов ошибок
- CI: контракт, линтер, тесты с -race, интеграционные на живом PostgreSQL,
  сборка стека со smoke, рендер диаграмм
- README: архитектура, схема БД, технические решения, упрощения MVP
- три проектных скилла: запуск стека, contract-first цикл, миграции

+++ internal/app/fakes_test.go
@@ -181,7 +181,7 @@ func (m *fakeMenuRepo) PublishVersion(_ context.Context, restaurantID int64, pro
 	version := prev.Menu.Version + 1
 
 	stored := make([]domain.Product, len(products))
-	nextID := int64(1000*restaurantID) + int64(version)*100
+	nextID := 1000*restaurantID + int64(version)*100
 	for i, p := range products {
 		p.ID = nextID + int64(i)
 		p.MenuID = int64(version)
